requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile 
#
argh==0.26.2              # via watchdog
astroid==1.6.5            # via pylint, pylint-celery, pylint-flask, requirements-detector
backports.functools-lru-cache==1.5  # via astroid, isort, pylint
click==7.0                # via pip-tools
colorama==0.4.1           # via pytest-watch
configparser==3.7.3       # via pydocstyle, pylint
docopt==0.6.2             # via pytest-watch
dodgy==0.1.9              # via prospector
enum34==1.1.6             # via astroid
futures==3.2.0            # via isort
github-backup==0.9.0
isort==4.3.10             # via pylint
jinja2==2.8
lazy-object-proxy==1.3.1  # via astroid
markupsafe==0.23          # via jinja2
mccabe==0.6.1             # via prospector, pylint
pathtools==0.1.2          # via watchdog
pep8-naming==0.4.1        # via prospector
pip-tools==3.4.0
prospector==1.1.6.2
py==1.4.31                # via pytest
pycodestyle==2.4.0        # via prospector
pydocstyle==3.0.0         # via prospector
pyflakes==1.6.0           # via prospector
pygithub==1.26.0
pylint-celery==0.3        # via prospector
pylint-django==0.8.1      # via prospector
pylint-flask==0.5         # via prospector
pylint-plugin-utils==0.5  # via prospector, pylint-celery, pylint-django, pylint-flask
pylint==1.9.4             # via prospector, pylint-celery, pylint-django, pylint-flask, pylint-plugin-utils
pytest-watch==4.2.0
pytest==2.9.1
pyyaml==5.1b3
requirements-detector==0.6  # via prospector
setoptconf==0.2.0         # via prospector
singledispatch==3.4.0.3   # via astroid, pylint
six==1.12.0               # via astroid, pip-tools, pydocstyle, pylint, singledispatch
snowballstemmer==1.2.1    # via pydocstyle
watchdog==0.9.0           # via pytest-watch
wrapt==1.11.1             # via astroid
